refactor(ui): extract useConnectSavedConnection from ConnectionList
Move the ~150-line credential-aware connect flow out of ConnectionList into
a reusable useConnectSavedConnection hook with no behavior change, so the
command palette (#1484) can reuse the exact same connect path. ConnectionList
now calls the hook; its insecure-FTP warning wrapper is unchanged.
